requirements.txt
# transformers
# torch
pandas
# accelerate>=0.26.0
openai
python-dotenv